Log when deleting old files because of space limit
git-svn-id: file:///D/data/svnrepos/uimager/trunk@1514 15f0fa91-3113-114a-a8b1-9261d5763783

+++ ToolBarChildFrm.cpp
@@ -997,7 +997,7 @@ void CVideoDeviceChildFrame::OnClose()
 				{
 					sMsg.Format(_T("%s stopping (running for %I64dday%s, %dhour%s, %dmin and %dsec)"),
 						pDoc->GetAssignedDeviceName(),
-						(LONGLONG)TimeSpan.GetDays(),
+						TimeSpan.GetDays(),
 						(TimeSpan.GetDays() == 1) ? _T("") : _T("s"),
 						TimeSpan.GetHours(),
 						(TimeSpan.GetHours() == 1) ? _T("") : _T("s"),
